+++ packages/runtime/src/types.ts
@@ -146,6 +146,7 @@ export interface Context {
 	gtxs: EncTx[]; // all transactions
 	args?: Uint8Array[];
 	debugStack?: number; //  max number of top elements from the stack to print after each opcode execution.
+	budget: number;
 	pooledApplCost: number; // total opcode cost for each application call for single/group tx
 	// inner transaction props
 	isInnerTx: boolean; // true if "ctx" is switched to an inner transaction