+++ packages/runtime/test/integration/gas-calculator.ts
@@ -41,6 +41,6 @@ describe("TEALv5: Pooled Opcode Cost calculation", function () {
 
 	it("Gas should be number", function () {
 		const receipt = runtime.executeTx([appCallParam])[0] as rtypes.AppInfo;
-		expect(receipt.gas).to.equal(98);
+		expect(receipt.gas).to.equal(99);
 	});
 });